Fix 64 bit conversion issue

+++ src/newt_core/NewtFile.c
@@ -434,7 +434,7 @@ newtRef NsLoadLib(newtRefArg rcvr, newtRefArg r)
 
     if (lib != NULL)
 	{
-        return NewtMakeInteger((int32_t)lib);
+        return NewtMakeInteger((intptr_t)lib);
 	}
 	else
 	{
